Widen activations from int32 to int64
Neuron.Activation, Baseline, Threshold, and PostFireReset are now int64,
matching the weight precision from the previous refactor. This removes
the int32 bottleneck in the activation domain that was identified as
the source of the 1% accuracy gap between float64 training and integer
inference.

Key changes:
- MaxActivation/MinActivation constants: int32 -> int64
- Neuron.Stimulate accepts int64 weight directly (no clamping)
- Network.Stimulate/Tick no longer clamp to int32 range
- Merged ClampAdd32 into ClampAdd (both domains now int64)
- Decay calculation uses native int64 arithmetic
- Updated all benchmarks, learning rules, and tests

29 files changed, consistent with the weight refactor scope.

+++ benchmark/fashionmnist/fashionmnist_deep_test.go
@@ -42,7 +42,7 @@ func TestFashionMNISTDeep(t *testing.T) {
 
 	intScale := float64(1 << 20)
 
-	intThreshold := int32(threshold * intScale)
+	intThreshold := int64(threshold * intScale)
 	net := bio.NewNetwork(uint32(total), 0, intThreshold, decayRate, 3)
 	net.LearningRule = bio.NoOpLearning{}
