+++ benchmark/mnist/mnist_early_exit_test.go
@@ -38,7 +38,7 @@ func TestMNISTEarlyExit(t *testing.T) {
 
 	intScale := float64(1 << 20)
 
-	intThreshold := int32(threshold * intScale)
+	intThreshold := int64(threshold * intScale)
 	net := bio.NewNetwork(uint32(total), 0, intThreshold, decayRate, 3)
 	net.LearningRule = bio.NoOpLearning{}
